+++ src/p_inter.c
@@ -259,12 +259,22 @@ void P_GiveCard(player_t *player, card_t card)
 {
   if (player->cards[card])
     return;
+    
   // [Nugget] Fix for "key pickup resets palette"
   if (STRICTMODE(!nugget_comp[comp_keypal]))
   { player->bonuscount += BONUSADD; }
   else
   { player->bonuscount = BONUSADD; }
+  
   player->cards[card] = 1;
+  
+  // [Nugget] Key pickup timer
+  if (STRICTMODE(timer_key_pickup == 2 || (timer_key_pickup && (demorecording||demoplayback))))
+  {
+    player->event_type = TIMER_KEYPICKUP;
+    player->event_time = leveltime;
+    player->event_tics = 5*TICRATE/2; // [crispy] 2.5 seconds
+  }
 }
 
 //